example/162_genai/pyproject.toml
[project]
name = "embed"
version = "0.1.0"
description = "Experiment with google genai"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "sqlite-minutils",
    #"google-genai[aoihttp]",
    "google-genai",
    "numpy",
    "pandas",
    "loguru",
]

[project.optional-dependencies]
test = [
    "pytest",
]
